specified correct data type

+++ src/main/java/com/company/RssReaderBot/db/repositories/RssFeedRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RssFeedRepository extends JpaRepository<RssFeed, Long> {
+public interface RssFeedRepository extends JpaRepository<RssFeed, Integer> {
 
     Optional<RssFeed> findByUserUserid(Long userid);
